+++ src/main/java/com/phion/spring/ioc/BeanDefinition.java
@@ -66,4 +66,14 @@ public class BeanDefinition {
     public void setPropertyValues(PropertyValues propertyValues) {
         this.propertyValues = propertyValues;
     }
+
+    @Override
+    public String toString() {
+        return "BeanDefinition{" +
+                "bean=" + bean +
+                ", beanClass=" + beanClass +
+                ", beanClassName='" + beanClassName + '\'' +
+                ", propertyValues=" + propertyValues +
+                '}';
+    }
 }